+++ public/css/styles.css
@@ -14,6 +14,12 @@ html {
     background-attachment: fixed;
     background-size: cover;
 }
+@media screen and (max-width: 768px) {
+    .has-bg-image {
+        background-attachment: scroll; /* Change to scroll for better mobile experience */
+        background-position: center; /* Adjust the position if needed */
+    }
+}
 .bulma-image {
     width: 200px;
 }
@@ -45,13 +51,24 @@ html {
 }
 
 #tsparticles {
-    position: fixed;
-    background-repeat: no-repeat !important;
-    /*background-size: cover !important; */
+    position: absolute;
+    /*background-repeat: no-repeat !important;*/
+    background-size: cover !important;
+    top: 0;
+    left: 0;
     width: 100%;
     height: 100%;
     z-index: 0;
   }
+
+.particle {
+    position: relative;
+    background-repeat: no-repeat !important;
+    background-size: cover !important;
+    width: 100%;
+    height: 100%;
+    z-index: 0;
+}  
 
 .pos-rel {
     position: relative;
